+++ R/print.R
@@ -181,6 +181,15 @@ print_time_varying <- function(x, m = 1) {
     colnames(mm_tv) <- c("coef.est", "coef.se")
     time_slices <- x$time_lu$time_from_data
     row.names(mm_tv) <- paste(rep(tv_names, each = length(time_slices)), time_slices, sep = "-")
+
+    # append AR(1) parameters if they are estimated
+    p <- tidy(x, effects = "ran_pars", model = m, silent = TRUE)
+    if(any(p$term == "rho_time")) {
+      rho_tv <- as.matrix(p[p$term=="rho_time",c("estimate","std.error")])
+      colnames(rho_tv) <- colnames(mm_tv)
+      row.names(rho_tv) <- paste("rho", tv_names, sep = "-")
+      mm_tv <- rbind(mm_tv, rho_tv)
+    }
   } else {
     mm_tv <- NULL
   }